Move solver fallback into fine package instead of root conftest
The Gurobi license detection and STANDARD_SOLVER override now lives in
ImplementedSolvers.set_standard_solver() (fine/utils.py), called from
fine/__init__.py on import. This ensures every consumer — unit tests,
notebook examples, and user scripts — gets the correct solver without
relying on pytest conftest.py discovery.
